requirements.txt
google-cloud-ndb>=1.6,<1.7
